Support floating point typesOnly for real, single and double. Expressions whose exponent doesn't have specified unary operator (+ or -) are not working

+++ PascalCompiler/src/pascal/language/nodes/arithmetic/NegationNode.java
@@ -15,7 +15,7 @@ public abstract class NegationNode extends UnaryNode{
     }
 	
 	@Specialization(rewriteOn = ArithmeticException.class)
-    protected int neg(int val) {
+    protected double neg(double val) {
 		return -val;
     }
 }
